+++ src/visualisers/solar-wavelength/demo/index.ts
@@ -1,0 +1,2 @@
+export { default } from './SolarDemo';
+export { default as SolarDemo } from './SolarDemo';